requirements.txt
# ---- 1) Install core PyTorch w/ CUDA ----
!pip install torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118

# ---- 2) Install all required pip packages ----
!pip install \
black==23.7.0 \
chardet==5.1.0 \
einops>=0.6.1 \
fairscale>=0.4.13 \
fire>=0.5.0 \
fsspec>=2023.6.0 \
invisible-watermark>=0.2.0 \
kornia==0.6.9 \
matplotlib>=3.7.2 \
natsort>=8.4.0 \
ninja>=1.11.1 \
numpy==1.24.4 \
pyarrow==17.0.0 \
bitsandbytes==0.43.2 \
open-clip-torch>=2.20.0 \
opencv-python==4.6.0.66 \
pandas>=2.0.3 \
pillow>=9.5.0 \
pyyaml>=6.0.1 \
scipy>=1.10.1 \
tensorboardx==2.6 \
timm>=0.9.2 \
tokenizers==0.12.1 \
tqdm>=4.65.0 \
transformers==4.30.0 \
triton==2.0.0 \
urllib3>=1.25.4,<1.27 \
webdataset>=0.2.33 \
wheel>=0.41.0 \
xformers==0.0.20 \
streamlit==1.31.0 \
streamlit-keyup==0.2.0 \
albumentations==0.4.3 \
pudb==2019.2 \
imageio==2.9.0 \
imageio-ffmpeg==0.4.2 \
omegaconf==2.1.1 \
wandb==0.15.4 \
h5py \
huggingface_hub \
datasets==3.0.2

# ---- 3) PyTorch Lightning (replace old 1.5.0) ----
!pip install pytorch-lightning==2.1.3

# ---- 4) Install Git-based repos ----
# CLIP
!pip install git+https://github.com/openai/CLIP.git@main

# taming-transformers variant
!pip install git+https://github.com/CompVis/taming-transformers.git

# ---- 5) Install the project itself, if setup.py/pyproject is present ----
!pip install -e .
